backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
feedparser==6.0.10
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
requests==2.31.0

